+++ lib/models/money_objects/accounts/account.dart
@@ -1,7 +1,7 @@
 // Imports
 import 'package:flutter/material.dart';
-import 'package:money/helpers/json_helper.dart';
 import 'package:money/helpers/string_helper.dart';
+import 'package:money/models/data_io/data.dart';
 import 'package:money/models/money_objects/accounts/account_types.dart';
 import 'package:money/models/money_objects/money_object.dart';
 import 'package:money/widgets/table_view/list_item_card.dart';
@@ -98,13 +98,23 @@ class Account extends MoneyObject<Account> {
   /// Currency
   /// 7|Currency|nchar(3)|0||0
   Field<Account, String> currency = Field<Account, String>(
+    type: FieldType.widget,
     importance: 4,
     name: 'Currency',
     serializeName: 'Currency',
     align: TextAlign.center,
     columnWidth: ColumnWidth.small,
     defaultValue: '',
-    valueFromInstance: (final Account instance) => instance.currency.value,
+    useAsDetailPanels: true,
+    valueFromInstance: (final Account instance) {
+      String locale = Data().currencies.fromSymbolToCountryAlpha2(instance.currency.value);
+      String flagName = getCountryFromLocale(locale).toLowerCase();
+      return Row(mainAxisAlignment: MainAxisAlignment.center, children: [
+        Image.asset('assets/flags/$flagName.png', height: 10),
+        const SizedBox(width: 4),
+        Text(instance.currency.value),
+      ]);
+    },
     valueForSerialization: (final Account instance) => instance.currency.value,
   );